Fix apple write offsets
For apple, we add four bytes at the BEGINNING of the packet. For that
first we create a byte vector of 4 bytes, and extend that by the existing
data, and then use the newly extended vector to do the libc::write .. The
existing data is at offset data.headroom into the existing buffer, but the
newly created buffer does NOT need headroom offset, its at offset 0.